+++ CHANGELOG.md
@@ -5,10 +5,11 @@
 - bug fixes
     - widgets: circular value feedback issues
     - editor: `ctrl + e` not working when hitm multiple times without releasing `ctrl`
-    - mobile: "prevent sleep" feature not working 
+    - mobile: "prevent sleep" feature not working
 
 - widgets
     - patchbay: add `in`, `out` and `both` modes to `exclusive` property
+    - panel: replace `verticalTabs` property with `tabsPosition` (supports 4 directions)
 
 - custom module
     - add optional `errorCallback` argument to`loadJSON` and `saveJSON`